remove update from builder component to useShex hook
