+++ src/lib/hooks/useShex.js
@@ -3,7 +3,17 @@ import data from '@solid/query-ldflex';
 import shexParser from '@shexjs/parser';
 import shexCore from '@shexjs/core';
 import unique from 'unique-string';
-import { findAnnotation, SolidError, fetchLdflexDocument, fetchSchema } from "@utils";
+import ldflex from "@solid/query-ldflex";
+import { namedNode } from "@rdfjs/data-model";
+import {
+  findAnnotation,
+  SolidError,
+  fetchLdflexDocument,
+  fetchSchema,
+  shexParentLinkOnDropDowns,
+  solidResponse,
+  ShexFormValidator
+} from "@utils";
 import { Expression, Annotation, ShexJ, Shape } from "@entities";
 
 type Options = {
@@ -14,6 +24,7 @@ type Options = {
 export const useShex = (fileShex: String, documentUri: String, rootShape: String, errorCallback: ?() => void) => {
     const [shexData, setShexData] = useState({});
     const [shexError, setShexError] = useState(null);
+    const [formValues, setFormValues] = useState({});
     let shapes = [];
     let seed = 1;
 
@@ -431,6 +442,200 @@ export const useShex = (fileShex: String, documentUri: String, rootShape: String
         }
     });
 
+    const onChange = e => {
+        const { value, name } = e.target;
+        const defaultValue = e.target.getAttribute('data-default');
+        const action = defaultValue === '' ? 'create' : value === '' ? 'delete' : 'update';
+
+        const data = {
+            [e.target.name]: {
+                value,
+                name,
+                action,
+                defaultValue,
+                predicate: e.target.getAttribute('data-predicate'),
+                subject: e.target.getAttribute('data-subject'),
+                prefix: e.target.getAttribute('data-prefix'),
+                parentPredicate: e.target.getAttribute('data-parent-predicate'),
+                parentSubject: e.target.getAttribute('data-parent-subject'),
+                valueExpr: JSON.parse(e.target.getAttribute('data-valuexpr')),
+                parentName: e.target.getAttribute('data-parent-name')
+            }
+        };
+        setFormValues({ ...formValues, ...data });
+    };
+
+    const _create = async field => {
+        const { subject, predicate, value } = field;
+        if (field.parentSubject) await _createLink(field);
+        await ldflex[subject][predicate].add(value);
+    };
+
+    const _createLink = async field => {
+        const { subject, parentPredicate, parentSubject } = field;
+        const id = `#${subject.split("#").pop()}`;
+        await ldflex[parentSubject][parentPredicate].add(namedNode(id));
+    };
+
+    const _setFieldValue = (value: String, prefix: ?String) =>
+        prefix ? namedNode(`${prefix}${value}`) : value;
+
+    const _deleteLink = async shexj => {
+        const { parentSubject, value: subject, parentPredicate } = shexj._formFocus;
+        const expressions = shexj.expression.expressions;
+        const { _formFocus } = shexj;
+        try {
+            if (_formFocus && !_formFocus.isNew) {
+                for (let expression of expressions) {
+                    const value = await ldflex[subject][expression.predicate];
+                    if (value) await ldflex[subject][expression.predicate].delete();
+                }
+                await ldflex[parentSubject][parentPredicate].delete(ldflex[subject]);
+            }
+        } catch (e) {
+            throw e;
+        }
+    };
+
+    const onDelete = async (shexj: ShexJ, parent: any = false) => {
+        try {
+            const { _formFocus } = shexj;
+            const { parentSubject, name, value, isNew } = _formFocus;
+            if (_formFocus && !isNew) {
+                if (shexParentLinkOnDropDowns(parent, shexj)) {
+                    await _deleteLink(shexj);
+                } else {
+                    const predicate = shexj.predicate || shexj._formFocus.parentPredicate;
+
+                    await ldflex[parentSubject][predicate].delete(value);
+                }
+            }
+            // Delete field from formValues object
+            const { [name]: omit, ...res } = formValues;
+            setFormValues(res);
+
+            return { status: 200, message: 'Form submitted successfully', fieldName: name};
+        } catch (error) {
+            let solidError = error;
+
+            if (!error.status && !error.code ) {
+                solidError = new SolidError(error.message, 'Ldflex Error', 500);
+            }
+            return solidError;
+        }
+    };
+
+    const onReset = () => {
+        setFormValues({});
+    }
+
+    const saveForm = async (key: String) => {
+        try {
+            let value = formValues[key].value;
+            let defaultValue = formValues[key].defaultValue;
+
+            // @TODO: find a better way to see if value is a predicate.
+            if (value.includes('#')) {
+                value = namedNode(value);
+                defaultValue = namedNode(defaultValue);
+            }
+
+            const field = {
+                ...formValues[key],
+                value: _setFieldValue(
+                    value,
+                    formValues[key].prefix
+                ),
+                defaultValue: _setFieldValue(
+                    defaultValue,
+                    formValues[key].prefix
+                )
+            };
+            switch (field.action) {
+                case "update":
+                    await ldflex[field.subject][field.predicate].replace(
+                        field.defaultValue,
+                        field.value
+                    );
+                    break;
+                case "create":
+                    await _create(field);
+                    break;
+                case "delete":
+                    await ldflex[field.subject][field.predicate].delete(
+                        field.defaultValue
+                    );
+                    break;
+                default:
+                    break;
+            }
+
+            // If save field was successful we update expression and parentExpression.
+            updateExpression(key);
+
+            return solidResponse(200, 'Form submitted successfully');
+
+        } catch (error) {
+            return error;
+        }
+    };
+
+
+    const onSubmit = async (e: Event) => {
+        try {
+            if (!documentUri || documentUri === "") {
+                throw Error("Document Uri is required");
+            }
+            e.preventDefault();
+            const validator = new ShexFormValidator(formValues);
+            const { isValid, updatedFields } = validator.validate();
+            const keys = Object.keys(formValues);
+
+            if (isValid && keys.length > 0) {
+
+                for await (const key of keys) {
+                    const result = await saveForm(key);
+
+                    if (result.code !== 200) {
+                        throw new SolidError(result.message, 'Error saving on Pod');
+                    }
+                }
+
+                setFormValues({});
+
+                return solidResponse(200, 'Form submitted successfully');
+
+            } else {
+                setFormValues({...updatedFields});
+
+                if (keys.length !== 0) {
+                    throw new SolidError('Please ensure all values are in a proper format.', 'ShexForm', 406);
+                }
+            }
+
+        } catch (error) {
+            let solidError = error;
+
+            if (!error.status && !error.code ) {
+                solidError = new SolidError(error.message, 'Ldflex Error', 500);
+            }
+            return solidError;
+        }
+    };
+
+    const updateExpression = (key: String) => {
+        const { name, parentName } = formValues[key];
+
+        updateShexJ(name, "update", {
+            isNew: false,
+            value: formValues[key].value
+        });
+
+        updateShexJ(parentName, "update", {
+            isNew: false
+        });
+    }
+
     useEffect(() => {
         toShexJForm();
 
@@ -440,6 +645,10 @@ export const useShex = (fileShex: String, documentUri: String, rootShape: String
         shexData,
         shexError,
         addNewShexField,
-        updateShexJ
+        updateShexJ,
+        formValues,
+        onSubmit,
+        onDelete,
+        onChange
     };
 };
